enable parallel jobs in feature extraction; minor changes on formatting
